requirements.txt
amqp==5.1.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
deprecated==1.2.13
h11==0.13.0
idna==3.3
kombu==5.2.4
outcome==1.2.0
packaging==21.3
prompt-toolkit==3.0.30
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pyopenssl==22.0.0
pyparsing==3.0.9
pysocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
rsa==4.9
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
telethon==1.24.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.8.3
wrapt==1.14.1
wsproto==1.1.0
